Initial work towards a reorganization of configuration
