+++ core/configuration/operations.py
