+++ core/models/__init__.py
@@ -28,10 +28,3 @@ from .process import (
     ProcessType,
     ProcessCategory,
 )
-from .configuration import (
-    AppConfigurationDefault,
-    get_configuration_default,
-    get_configuration_choices,
-    list_configuration_keys,
-    InstanceConfiguration,
-)
